Implement .pth file approach for vLLM multi-API server patching
## Changes

**Core Implementation:**
- Create patch.py with idempotent apply_patches() function that patches:
  - AsyncEngineArgs.create_engine_config to inject worker_extension_cls
  - build_app to add custom /update_weights and /reload_weights endpoints
- Add prime_rl_vllm_patch.pth to auto-apply patches on Python startup
- Simplify server.py (removed fork enforcement and inline patching)

**Build System:**
- Configure hatchling to install .pth file in site-packages via shared-data
- Update requires-python to >=3.10

**Key Benefits:**
- Patches apply in ALL spawned workers (parent + child processes)
- Works with any multiprocessing method (fork, spawn, forkserver)
- Platform-agnostic (no fork() dependency)
- Cleaner code: 62 lines added, 72 lines removed

**Testing:**
Comprehensive testing on production hardware (2x NVIDIA H100):
- ✅ Single-API server mode (--api-server-count 1)
- ✅ Multi-API server mode (--api-server-count 2)
- ✅ Custom endpoints: 100% success rate across all API servers
- ✅ Worker extension: CheckpointWorker successfully injected
- ✅ collective_rpc: update_weights routes to worker correctly
