+++ src/prime_rl/data/prime_rl_vllm_patch.pth
@@ -1,0 +1,1 @@
+import prime_rl.inference.vllm.patch; prime_rl.inference.vllm.patch.apply_patches()
